rename the 'other' folder to 'util'
