+++ src/js/util/getFirebaseApp.js
